Corrections to heapUp and heapDown
Added conditional statements to heapUp and heapDown to
prevent error and/or infinite recursion.

For heapUp check if the current index is greater than 0 to
prevent a negative / out of bounds parent.

For heapDown only attempt if there is at least a left child within
the bounds of the current size, and only compare left and right
children if there is a right child within the bounds of the
current size.